realtime/requirements.txt
websockets>=12.0
numpy>=1.26.0
sounddevice>=0.4.6
requests>=2.32.0
google-cloud-speech>=2.27.0
google-cloud-texttospeech>=2.16.5
webrtcvad>=2.0.10
